Passing metadata to void nil panic in the providers
