+++ backend/internal/flow/executor/attribute_collector.go
@@ -136,8 +136,9 @@ func (a *attributeCollector) HasRequiredInputs(ctx *providers.NodeContext,
 		return true
 	}
 
+	metadata := core.BuildGetAttributesMetadata(ctx)
 	// Update the executor response with the required inputs retrieved from authenticated user attributes.
-	authUser, authnUserAttrs, svcErr := a.authnProvider.GetUserAttributes(ctx.Context, nil, nil, execResp.AuthUser)
+	authUser, authnUserAttrs, svcErr := a.authnProvider.GetUserAttributes(ctx.Context, nil, metadata, execResp.AuthUser)
 	if svcErr != nil {
 		logger.Warn(ctx.Context, "Failed to retrieve authenticated user attributes")
 	}
