+++ backend/internal/flow/executor/consent_executor.go
@@ -157,7 +157,7 @@ func (e *consentExecutor) checkConsent(ctx *providers.NodeContext, execResp *pro
 	promptData, svcErr := e.consentEnforcer.ResolveConsent(
 		ctx.Context, ouID, appID, appName, entityRef.EntityID,
 		essentialAttributes, optionalAttributes, authorizedPermissions,
-		availableAttributes, forceReprompt, buildRuntimeMetadata(ctx))
+		availableAttributes, forceReprompt, core.BuildRuntimeMetadata(ctx))
 	if svcErr != nil {
 		if svcErr.Type == tidcommon.ClientErrorType {
 			logger.Debug(ctx.Context, "Client error while resolving user consent", log.Any("error", svcErr))
@@ -261,7 +261,7 @@ func (e *consentExecutor) handleConsentDecisions(ctx *providers.NodeContext, exe
 	// Always record consent decisions (including denials) for audit/compliance purposes.
 	// The session token is used to verify completeness and enforce essential attribute rules
 	consentRecord, svcErr := e.consentEnforcer.RecordConsent(ctx.Context, ouID, appID, userID,
-		&decisions, sessionToken, validityPeriod, buildRuntimeMetadata(ctx))
+		&decisions, sessionToken, validityPeriod, core.BuildRuntimeMetadata(ctx))
 	if svcErr != nil {
 		// Essential consent denied: the consent record was persisted but the user denied
 		// a required attribute, so the flow cannot proceed
