+++ backend/internal/flow/executor/credentials_auth_executor.go
@@ -203,7 +203,7 @@ func (b *credentialsAuthExecutor) authenticateUser(ctx *providers.NodeContext,
 	}
 
 	// For authentication flows, call Authenticate directly.
-	metadata := buildAuthnMetadata(ctx)
+	metadata := core.BuildAuthnMetadata(ctx)
 	authUser, authenticatedClaims, svcErr := b.authnProvider.AuthenticateUser(ctx.Context, userIdentifiers,
 		userCredentials, nil, metadata, execResp.AuthUser)
 	execResp.AuthUser = authUser
